Add todo about figuring out how to handle environment variables
When Tanzawa is running via uWSGI / docker-compose, we probably want to
to pass the environment variables to Tanzawa via Docker-compose and not
have it read directly from a .env file.

What changes (if any) will this require when running Tanzawa directly
where we specify the ENV_FILE when running the command / in the env.