Fetch resized image given just the mime
When an image is over the "max" size served by Tanzawa publicly, Tanzawa automatically creates a resized version and saves it to disk.

The next time we request the image, if the original image is too large, Tanzawa should return resized image.

However, when querying for an existing resized image (so we do not have to perform the resize operation again), we only looked for the image's original size.

Change our logic to look for the original image, then the maximum resized image, then finally create a new processed file.

Fixes PYTHON-DJANGO-B