Use a custom spatialite database wrapper
There's an incompatibility between Django 4.2 and spatialite that results in an error where it tries to re-initialize the metadata, even if it's already been initialized.

The error message is:

```
django.db.utils.OperationalError: error in trigger ISO_metadata_reference_row_id_value_insert: no such column: rowid
```

refs: https://code.djangoproject.com/ticket/32935

Using a custom wrapper that skips the initialization fixes the issue. I was trying to avoid using a custom wrapper for months in hopes that a patch would land in the libraries.

The bug has made developing Tanzawa no fun and effectively pause. Now I can get on with development again. Yay!