Pass timezone to arrow when getting activity time
Start date local has a Z on it, which implies it's UTC, however it's actually localtime (as the name implies).

Using .to(timezone) results in the datetime being adjusted, and when you're ahead of GMT, means posts won't be accessible until UTC catches up with localtime. So for Japan this is 9 hours later.

Pass in the timezone directly so the resulting datetime is localtime with a timezone.