Ensure bookmarks can filter multi-word tags
Before this commit filtering bookmarks by tags with a space in them e.g. "climate change" would search for items tagged "climate" and "change".

After this commit we quote all tags that are passed in so in the example above we'd search for "climate change".

This also has the side benefit of allowing us to pass _multiple_ tags in and Tanzawa filters by both tags.